[FIX] Rename lib files in bots
Renaming non-loadable bot files, so it wont get loaded by the dynamic
bots list.
